Create a new runtimeusers group
This group can be used in the build phase of a Dockerfile to grant the runtime user extra permissions on files / directories.

Due to https://github.com/docker/for-linux/issues/388, recursive chown during the container initialization (or during container runtime) is extremely slow. In this way, the change of ownership can be done at build time, and then the runtime user can be assigned to this new group in order to have extended access to these resources.